+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [1.3.5]
 - Fixes cases where a requests were processed twice from cache
 - Multiple On User Login requests can be fired consequtively.
+- Dynamic image dimensions for web carousel
 
 ## [1.3.4] - 1 Dec, 2022
 - Fixes cases where multiple request can be fired without a gcookie.